Experiment with per-project spellchecker
I generally code with `set nospell`, turning on `spell` occasionally to
spot-check my work (accepting the fact that Vim's spellchecker will flag
a lot of code values and company-specific terminology as typos).

I'm going to try out building a dictionary of expected values for each
project so that `set spell` gives useful output.

In each project where I'm trying this, I will have a project-specific
`.vimrc` that looks like so:

    set spell  " Turn spelling on (it's off by default in my global .vimrc)
    set spelllang=en_us
    set spellfile=.spellfile.en.utf-8.add  " Special filename, Git ignores

    " http://andrew.stwrt.ca/posts/project-specific-vimrc/
    set secure  " Disable unsafe commands.

In order to add a word to the project's list of spelling exceptions,
just press `zg` while hovering the word.

This is based off the recommended setup by Thoughtbot:
https://thoughtbot.com/blog/opt-in-project-specific-vim-spell-checking-and-word-completion

We'll see how this goes!