Version control dotfiles backups
Previous iterations of this script had a nasty side effect: Running it
twice could accidentally remove the original versions of files!

Run 1:
- Copy existing content to ~/dotfiles_old
- Replace original dotfiles in ~/

Run 2:
- Copy dotfiles to ~/dotfiles_old, overwriting saved files
- Notice that the originals are now gone!

Solve this problem by version-controlling backups with a Git repository.
This provides a simple, effective way of logging a history.

As a benefit, this also solves the problem of accidentally nesting
symlinks in vimfiles.