No need to pass `-r` or `-f` to `rm`
The `-L` check confirms that `~/.vim` is a symlink (either to a file, or
to a directory). There's no need to use the `-r` flag, since symlinks
can always be deleted without it (even if they point to directories).
Additionally, the `-f` flag is not necessary, since we know the file
exists. Furthermore, if permissions prevent a normal removal, we don't
want an automated script getting in the way of that.