Only initialize pyenv on systems with it
I use pyenv for local development, but it's generally not present on
remote machines. This is fine - I'd like to be able to use all my other
tools without my bashrc failing to source. Only initialize pyenv if it's
actually installed.

(Also includes a related Vim commit for conditionally executing yarn -
both of these commits relate to installation on a fresh system)