Quick alias to globally install w/ pip
To force my pip environment to be clean,
I use `export PIP_REQUIRE_VIRTUALENV=true`.

This is super handy, and exactly what I want 99% of the time. However,
there is sometimes a reason to allow pip to install globally (or even
just to `--user`). Add a quick alias that does just that.