Handle special filenames in Vim + git grep aliases
I use `vgg` and `vgc` to quickly search for a regex in a repo and start
loading files that have matches. 99% of the time, I'm in a repository
with files that are just ASCII without spaces. However, to handle odd
filenames, using the NUL delimiter + `xargs` is a more sensible
approach.