Fix 24 bit/true color in tmux + Alacritty
It's important for tmux to be able to display full 24 bit colors in
order for tools like `git-delta` to be able to display properly.

This change is only needed to support working with Alacritty.
ITerm2 supports 24 bit colors by default, but we need this override in
tmux to make very sure that the surrounding TERM is told that it
supports 256 colors.