Let ripgrep through fzf search hidden files
Very often, I want to use Ctrl-T to find version-controlled files which
happen to be prefixed with a dot. It's a bit annoying that fzf/rg
excludes these files by default. Gitignore should already be covering
this case in general!

Granted, using Ctrl-T in a homedir (e.g. with a massive `.cache`) will
now be wildly inefficient, but that's fine -- I pretty much always use
fzf in a Git repo.